Extract ValidatorVerifier
