+++ packages/processor/source/verifiers/index.ts
@@ -2,4 +2,5 @@ export * from "./chained-verifier";
 export * from "./forged-transactions-verifier";
 export * from "./incompatible-transactions-verifier";
 export * from "./nonce-verifier";
+export * from "./validator-verifier";
 export * from "./verify-block-verifier";
