+++ packages/processor/source/verifiers/validator-verifier.ts
@@ -1,0 +1,86 @@
+import { inject, injectable } from "@mainsail/container";
+import { Contracts, Identifiers } from "@mainsail/contracts";
+import { Services, Utils } from "@mainsail/kernel";
+
+@injectable()
+export class ValidatorVerifier implements Contracts.BlockProcessor.Handler {
+	@inject(Identifiers.Application)
+	protected readonly app!: Contracts.Kernel.Application;
+
+	@inject(Identifiers.Cryptography.Configuration)
+	private readonly configuration: Contracts.Crypto.IConfiguration;
+
+	@inject(Identifiers.TriggerService)
+	private readonly triggers!: Services.Triggers.Triggers;
+
+	@inject(Identifiers.LogService)
+	private readonly logger!: Contracts.Kernel.Logger;
+
+	public async execute(roundState: Contracts.Consensus.IRoundState): Promise<boolean> {
+		const block = roundState.getProposal().toData().block;
+
+		const roundInfo: Contracts.Shared.RoundInfo = Utils.roundCalculator.calculateRound(
+			block.data.height,
+			this.configuration,
+		);
+
+		const validators: Contracts.State.Wallet[] = await this.triggers.call("getActiveValidators", {
+			roundInfo,
+		});
+
+		const forgingInfo: Contracts.Shared.ForgingInfo = await Utils.forgingInfoCalculator.calculateForgingInfo(
+			block.data.timestamp,
+			block.data.height,
+			this.app,
+		);
+
+		const forgingValidator: Contracts.State.Wallet = validators[forgingInfo.currentForger];
+
+		const walletRepository = this.app.getTagged<Contracts.State.WalletRepository>(
+			Identifiers.WalletRepository,
+			"state",
+			"blockchain",
+		);
+		const generatorWallet: Contracts.State.Wallet = await walletRepository.findByPublicKey(
+			block.data.generatorPublicKey,
+		);
+
+		let generatorUsername: string;
+		try {
+			generatorUsername = generatorWallet.getAttribute("validator.username");
+		} catch {
+			return false;
+		}
+
+		if (!forgingValidator) {
+			this.logger.debug(
+				`Could not decide if validator ${generatorUsername} (${
+					block.data.generatorPublicKey
+				}) is allowed to forge block ${block.data.height.toLocaleString()}`,
+			);
+		} else if (forgingValidator.getPublicKey() !== block.data.generatorPublicKey) {
+			Utils.assert.defined<string>(forgingValidator.getPublicKey());
+
+			const forgingWallet: Contracts.State.Wallet = await walletRepository.findByPublicKey(
+				forgingValidator.getPublicKey()!,
+			);
+			const forgingUsername: string = forgingWallet.getAttribute("validator.username");
+
+			this.logger.warning(
+				`Validator ${generatorUsername} (${
+					block.data.generatorPublicKey
+				}) not allowed to forge, should be ${forgingUsername} (${forgingValidator.getPublicKey()})`,
+			);
+
+			return false;
+		}
+
+		this.logger.debug(
+			`Validator ${generatorUsername} (${
+				block.data.generatorPublicKey
+			}) allowed to forge block ${block.data.height.toLocaleString()}`,
+		);
+
+		return true;
+	}
+}
